refactor songs page
